requirements.txt
python-telegram-bot
yfinance
pandas_ta
pandas
numpy
requests